+++ iexec-core-library/src/main/java/com/iexec/core/config/PublicConfiguration.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020-2024 IEXEC BLOCKCHAIN TECH
+ * Copyright 2020-2025 IEXEC BLOCKCHAIN TECH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -22,10 +22,7 @@ import lombok.Builder;
 import lombok.Value;
 
 /**
- * The {@link #blockchainAdapterUrl} and {@link #configServerUrl} fields are repetitive,
- * so if both are populated, they must contain the same value.
- * Duplicating these fields ensures worker/scheduler compatibility on different v8 versions.
- * In v9, the {@link #blockchainAdapterUrl} field will be removed.
+ * Configuration exposed by the scheduler and available publicly to all workers.
  */
 @Value
 @Builder
@@ -33,11 +30,6 @@ import lombok.Value;
 public class PublicConfiguration {
     String workerPoolAddress;
     String schedulerPublicAddress;
-    /**
-     * @deprecated Use {@link #configServerUrl} instead.
-     */
-    @Deprecated(forRemoval = true)
-    String blockchainAdapterUrl;
     String configServerUrl;
     String resultRepositoryURL;
     long askForReplicatePeriod;