+++ CHANGELOG.md
@@ -14,7 +14,7 @@ All notable changes to this project will be documented in this file.
 - Upgrade to Spring Boot 3.0.13. (#728)
 - Upgrade to Spring Cloud 2022.0.5. (#728)
 - Upgrade to Mongock 5.4.0. (#728)
-- Upgrade to `org.springdoc:springdoc-openapi-starter-webmvc-ui` 2.6.0. (#728)
+- Upgrade to Spring Doc OpenAPI 2.6.0. (#728)
 
 ## [[8.6.0]](https://github.com/iExecBlockchainComputing/iexec-core/releases/tag/v8.6.0) 2024-12-23