+++ CHANGELOG.md
@@ -9,6 +9,10 @@ All notable changes to this project will be documented in this file.
 - Configuration server is now optional by default. (#728)
 - Improve switch statements after Java 17 migration. (#729)
 
+### Breaking API changes
+
+- Remove deprecated blockhainAdapterUrl field from PublicConfiguration. (#729)
+
 ### Dependency Upgrades
 
 - Upgrade to `eclipse-temurin:17.0.13_11-jre-focal`. (#728)