+++ CHANGELOG.md
@@ -11,7 +11,8 @@ All notable changes to this project will be documented in this file.
 
 ### Breaking API changes
 
-- Remove deprecated blockhainAdapterUrl field from PublicConfiguration. (#729)
+- Remove deprecated blockhainAdapterUrl field from PublicConfiguration. (#730)
+- Remove deprecated REST endpoints in TaskController. (#731)
 
 ### Dependency Upgrades